+++ cmd/conduit/main.go
@@ -14,11 +14,11 @@ import (
 
 	"github.com/algorand/indexer/cmd/conduit/internal/list"
 	"github.com/algorand/indexer/conduit"
+	"github.com/algorand/indexer/conduit/loggers"
 	"github.com/algorand/indexer/conduit/pipeline"
 	_ "github.com/algorand/indexer/conduit/plugins/exporters/all"
 	_ "github.com/algorand/indexer/conduit/plugins/importers/all"
 	_ "github.com/algorand/indexer/conduit/plugins/processors/all"
-	"github.com/algorand/indexer/loggers"
 	"github.com/algorand/indexer/version"
 )
